feat(semaphoreci): embed project ref to pipeline
