+++ sources/semaphoreci-source/src/semaphoreci/semaphoreci.ts
@@ -5,11 +5,11 @@ import axios, {
   AxiosResponseHeaders,
 } from 'axios';
 import {AirbyteLogger} from 'faros-airbyte-cdk';
-import {DateTimeFormatOptions} from 'luxon';
 import parse, {Links} from 'parse-link-header';
-import {pipeline} from 'stream';
 import {Memoize} from 'typescript-memoize';
 import {VError} from 'verror';
+
+import {Pipeline, Project} from './models';
 
 const REST_API_VERSION = 'v1alpha';
 const SEMAPHORE_PAGE_HEADER = 'link';
@@ -19,60 +19,6 @@ export const UNAUTHORIZED_ERROR_MESSAGE =
   'SemaphoreCI API authorization failed. Verify that your API Token and Organization name are setup correctly';
 export const MISSING_OR_INVISIBLE_RESOURCE_ERROR_MESSAGE =
   'Resource not found or is not visible to the user';
-
-export interface ProjectSpec {
-  readonly visibility: string;
-  readonly name: string;
-  readonly owner: string;
-}
-
-export interface ProjectMeta {
-  readonly owner_id: string;
-  readonly org_id: string;
-  readonly name: string;
-  readonly id: string;
-  readonly description: string;
-}
-
-export interface Project {
-  readonly spec: ProjectSpec;
-  readonly metadata: ProjectMeta;
-}
-
-export interface Pipeline {
-  readonly terminate_request: string;
-  readonly queuing_at: string;
-  readonly working_directory: string;
-  readonly name: string;
-  readonly branch_id: string;
-  readonly project_id: string;
-  readonly running_at: string;
-  readonly partially_rerun_by: string;
-  readonly with_after_task: string;
-  readonly state: string;
-  readonly snapshot_id: string;
-  readonly commit_message: string;
-  readonly commit_sha: string;
-  readonly terminated_by: string;
-  readonly after_task_id: string;
-  readonly created_at: string;
-  readonly error_description: string;
-  readonly repository_id: string;
-  readonly yaml_file_name: string;
-  readonly pending_at: string;
-  readonly ppl_id: string;
-  readonly stopping_at: string;
-  readonly wf_id: string;
-  readonly done_at: string;
-  readonly result: string;
-  readonly compile_task_id: string;
-  readonly hook_id: string;
-  readonly branch_name: string;
-  readonly promotion_of: string;
-  readonly switch_id: string;
-  readonly result_reason: string;
-  readonly partial_rerun_of: string;
-}
 
 export interface SemaphoreCIConfig {
   readonly organization: string;
@@ -244,6 +190,8 @@ export class SemaphoreCI {
     const startTimeMax =
       startTime > this.startDate ? startTime : this.startDate;
 
+    const projects = await this.getProjects();
+
     const pipelines = await this.paginate<Pipeline>(
       ({nextPage}) =>
         this.restClient.get(
@@ -257,6 +205,11 @@ export class SemaphoreCI {
     );
 
     for (const pipeline of pipelines) {
+      const project = projects.find(
+        (project) => project.metadata.id === pipeline.project_id
+      );
+      pipeline.project = project;
+
       yield pipeline;
     }
   }
