+++ sources/semaphoreci-source/test/index.test.ts
@@ -207,10 +207,15 @@ describe('index', () => {
 
         SemaphoreCI.instance = jest.fn().mockImplementation(() =>
           makeSemaphoreCI({
-            get: restClient.mockResolvedValue({
-              data: readTestResourceFile('pipelines.json'),
-              headers: generateLinkHeaders(),
-            }),
+            get: restClient
+              .mockResolvedValueOnce({
+                data: readTestResourceFile('projects.json'),
+                headers: generateLinkHeaders(),
+              })
+              .mockResolvedValueOnce({
+                data: readTestResourceFile('pipelines.json'),
+                headers: generateLinkHeaders(),
+              }),
           })
         );
 
@@ -231,8 +236,10 @@ describe('index', () => {
           pipelines.push(pipeline);
         }
 
-        expect(restClient).toHaveBeenCalledTimes(1);
-        expect(restClient).toHaveBeenCalledWith(
+        expect(restClient).toHaveBeenCalledTimes(2);
+        expect(restClient).toHaveBeenNthCalledWith(1, 'projects');
+        expect(restClient).toHaveBeenNthCalledWith(
+          2,
           'pipelines?page=1&project_id=bea7e6ed-911e-4172-80f8-7ab58b541a86'
         );
         expect(pipelines).toStrictEqual(
@@ -246,10 +253,15 @@ describe('index', () => {
         SemaphoreCI.instance = jest.fn().mockImplementation(() =>
           makeSemaphoreCI(
             {
-              get: restClient.mockResolvedValue({
-                data: readTestResourceFile('pipelines.json'),
-                headers: generateLinkHeaders(),
-              }),
+              get: restClient
+                .mockResolvedValueOnce({
+                  data: readTestResourceFile('projects.json'),
+                  headers: generateLinkHeaders(),
+                })
+                .mockResolvedValueOnce({
+                  data: readTestResourceFile('pipelines.json'),
+                  headers: generateLinkHeaders(),
+                }),
             },
             [],
             new Date('2022-07-25T21:46:00Z')
@@ -273,8 +285,10 @@ describe('index', () => {
           pipelines.push(pipeline);
         }
 
-        expect(restClient).toHaveBeenCalledTimes(1);
-        expect(restClient).toHaveBeenCalledWith(
+        expect(restClient).toHaveBeenCalledTimes(2);
+        expect(restClient).toHaveBeenNthCalledWith(1, 'projects');
+        expect(restClient).toHaveBeenNthCalledWith(
+          2,
           'pipelines?page=1&project_id=bea7e6ed-911e-4172-80f8-7ab58b541a86'
         );
         expect(pipelines).toStrictEqual([
@@ -287,10 +301,15 @@ describe('index', () => {
 
         SemaphoreCI.instance = jest.fn().mockImplementation(() =>
           makeSemaphoreCI({
-            get: restClient.mockResolvedValue({
-              data: readTestResourceFile('pipelines.json'),
-              headers: generateLinkHeaders(),
-            }),
+            get: restClient
+              .mockResolvedValueOnce({
+                data: readTestResourceFile('projects.json'),
+                headers: generateLinkHeaders(),
+              })
+              .mockResolvedValueOnce({
+                data: readTestResourceFile('pipelines.json'),
+                headers: generateLinkHeaders(),
+              }),
           })
         );
 
@@ -311,8 +330,10 @@ describe('index', () => {
           pipelines.push(pipeline);
         }
 
-        expect(restClient).toHaveBeenCalledTimes(1);
-        expect(restClient).toHaveBeenCalledWith(
+        expect(restClient).toHaveBeenCalledTimes(2);
+        expect(restClient).toHaveBeenNthCalledWith(1, 'projects');
+        expect(restClient).toHaveBeenNthCalledWith(
+          2,
           'pipelines?page=1&project_id=bea7e6ed-911e-4172-80f8-7ab58b541a86&branch_name=main'
         );
         expect(pipelines).toStrictEqual(
@@ -323,6 +344,10 @@ describe('index', () => {
       test('full_refresh sync mode - paginated', async () => {
         const restClient = jest.fn();
         const mockGet = restClient
+          .mockResolvedValueOnce({
+            data: readTestResourceFile('projects.json'),
+            headers: generateLinkHeaders(),
+          })
           .mockResolvedValueOnce({
             data: readTestResourceFile('pipelines.json'),
             headers: generateLinkHeaders(1, 2, 2),
@@ -355,13 +380,14 @@ describe('index', () => {
           pipelines.push(pipeline);
         }
 
-        expect(restClient).toHaveBeenCalledTimes(2);
+        expect(restClient).toHaveBeenCalledTimes(3);
+        expect(restClient).toHaveBeenNthCalledWith(1, 'projects');
         expect(restClient).toHaveBeenNthCalledWith(
-          1,
+          2,
           'pipelines?page=1&project_id=bea7e6ed-911e-4172-80f8-7ab58b541a86'
         );
         expect(restClient).toHaveBeenNthCalledWith(
-          2,
+          3,
           'pipelines?page=2&project_id=bea7e6ed-911e-4172-80f8-7ab58b541a86'
         );
         expect(pipelines).toStrictEqual([
